feat(todo): add authentication to item/create

+++ todo/src/item/item.service.ts
@@ -10,15 +10,17 @@ import { ItemNotFoundException } from './exceptions'
 export class ItemService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createDto: CreateDto): Promise<Item> {
-    // TODO: Make sure listId belongs to the user
+  async create(userId: string, createDto: CreateDto): Promise<Item> {
     try {
       const item = await this.prisma.item.create({
         data: {
           title: createDto.title,
           list: {
             connect: {
-              id: createDto.listId,
+              userIndex: {
+                userId,
+                id: createDto.listId,
+              },
             },
           },
         },
@@ -34,9 +36,21 @@ export class ItemService {
     }
   }
 
-  async delete(id: string): Promise<void> {
-    // TODO: Make sure item belongs to the user
+  async delete(userId: string, id: string): Promise<void> {
     try {
+      const item = await this.prisma.item.findUniqueOrThrow({
+        where: {
+          id,
+        },
+        include: {
+          list: true,
+        },
+      })
+
+      if (item.list.userId !== userId) {
+        throw new ItemNotFoundException()
+      }
+
       await this.prisma.item.delete({
         where: {
           id,
@@ -44,7 +58,7 @@ export class ItemService {
       })
     } catch (error) {
       if (ErrorUtil.isNotFoundError(error)) {
-        throw new BadRequestException('Item not found')
+        throw new ItemNotFoundException()
       }
 
       throw error
